// planning-idsdp-rms-frontend/src/app/main/modules/rpm/models/ResearcherProposalRscWorkingInOrganization.ts
export class ResearcherProposalRscWorkingInOrganization {

    id: number;
    uuid: string;
    researcherProposalId: number;
    researcherProfileId: number;
    researcherName: string;
    telephoneNo: string;
    age: any;
    occupation: string;
    isMainResearcher: boolean;
    mobileNo: string;
    emailAddress: string;
    nidNo: string;
    designation: string;
    educationQualification: string;
    isDeleted: number = 0;
    researcherTypeName: string;
    personalDigitalInformation: string;
    fileName: string;
}