// planning-idsdp-rms-frontend/src/app/main/modules/rpm/models/ResearcherProposalInformation.ts
export class ResearcherProposalInformation {

    id: number;
    uuid: string;
    researcherProposalId: number;
    introduction: string;
    statementOfProblem: string;
    objectivesOfTheStudy: string;
    formulationResearch: string;
    conceptualFramework: string;
    reviewOfLiterature: string;
    rationaleOfTheStudy: string;
    scopeOfTheStudy: string;
    methodsOfTheStudy: string;
    expectedOutput: string;
    relationWithSocialPolicy: string;
    tentativeChapterization: string;
    actionPlanTentativeBudget: string;
    tentativeBudget:string;
    bibliographyReferences: string;
    additionalInfoTopicProposal: string;
    isEditable: 0 | 1;
}